Massive reformatting.
Decided it would be good to start strictly adhering to a specific
coding style.  And what better way to do that then use an
auto-formatting tool before committing?
I'm using the tool "Artistic Style" (http://astyle.sourceforge.net/).
It can be run with all the appropriate settings with the shell script
"format_code.sh".  It should be trivial to port that script to other
platforms as necessary.

Also added a TODO file, to help me keep track of tasks I can work on.