Added a sphere primitive.
This will be the go-to primitive for figuring out interfaces to
primitives.  It's super simple, and let's me play around without
much hassle.

It may stick around as a full-fledged primitive for production
rendering, even.  Although it will never support displacement mapping
since it is not a micropolygon-based surface.  But it could be useful
for e.g. particle systems.