Big refactor complete.
Rendering is now split up into three classes:
- Renderer
- Integrator
- Tracer

The Renderer is essentially the API for controlling rendering.  The
Integrator and Tracer are in turn used by the Renderer to actually
do the rendering.

There is also a Scene class, which contains the complete description
of a scene to be rendered.

In addition to the refactor, I also have:
- Added command-line options
- Added super stupid threading support, just as a test to get my feet wet
- Fixed the seeding of the RNG to be consistent
- Added a unit test for the RNG to guarantee that the fix stays