Implemented a simple disk cache class, and use that for output image buffers.
The DiscCache class behaves similarly to a c++ Vector, except that it
stores most of its data on disk, only holding a small portion in RAM
at a time.  This is ideal for large data sets that are typically
accessed very coherently.

I'm using this for render output image buffers to allow very high
resolution output images with numerous render layers to be handled
transparently whilst only taking up a very small amount of RAM.

Currently this is limited to single-threaded implementations.  However,
sample accumulation (which is the main task affected by this) is a very
small part of the over-all render time, so I'm not too worried as it's
hardly a bottleneck.  A multi-threaded implementation would very cool,
however, and might allow the use of DiskCache elsewhere in Psychopath
at some point.