Refactoring for multi-primitive ray testing.
The tracer now accumulates multiple potential intersecting primitives
at once, and then tests them each in sequence.  Currently only works
with the PrimArray collection type as proof-of-concept, as well as
to act as verification for getting the BVH working properly later
on.

This is a preparatory step for implementing ray reordering.