Implemented counting sort for ray reordering.
Counting sort sorts in O(N) time instead of O(N*log(N)) time,
so it scales much better to large numbers of potential intersections.
This is important because otherwise not only does RAM usage increases
with large numbers of rays shot at once, but also time per ray.  Using
counting sort helps make it an easy choice to use large numbers of rays
as long as you have the RAM for it.