Refactored sampling code to work on simple arrays.
This allows all of the samples to be kept in a contiguous chunk
in memory.  It doesn't appear to provide any speedup at the moment,
but it should simplify parallel sample generation in the future.

Also moved grid code to a new "usurface" directory.  This is in preparation
for some refactoring and new code.