Experimental rendering with leaf bboxes.
Inspired by the following papers:
- Floating-Point Precision Ray Tracing of Free-Form Surfaces
- Two-Level Ray Tracing with Reordering for Highly Complex Scenes

The idea is that for micropolygon grids, to test against the bboxes of
the micropolygons rather than the micropolys themselves.  The speedup
is minor, but it may have some other benefits, esp. e.g. known error
bounds..

Also beginning work on migrating things to a MicroSurface class,
which will be the low-level representation of surfaces, that all
ray testing happens against.