Improved RNG and fixed bug in the path tracing integrator.
The RNG class by default now uses a thread-safe incremented seed
when creating new RNG objects.  This means that unique RNG's are
guaranteed throughout the program unless the seed is manually set.

There was a bug in the path tracer that always spawned bounce rays
in the hemisphere of the surface normal, even when the incoming
ray came from the opposite side.  This was obviously incorrect, and
the visual effect of it was much brighter bounced light than
should have been the case.  It's interesting to note, however,
that the effect was visually pleasing and not obviously
incorrect.