Modified LRU cache to return shared_ptr's of the data.
This simplifies usage of the LRU cache significantly, because you don't
have to manage "closing" the data.  And that, in turn, will reduce buggy
usage of the LRU cache.

Also fixed bug that resulted in the weird bright pixels at the corners
of the pixel-blocks.  The path data structures weren't getting cleared
out correctly.