BVH4 now traverses the nearest child first.
The other children are not traversed in any particular order, but even
just making sure that the nearest child is traversed first avoids the
most pathalogical traversals.

In practice, this seems to slow down rendering of some scenes, while speeding
up rendering of others.  The theory is sound, though, and I strongly
suspect that it's an over-all win, even if only for the sake of avoiding
rare pathalogical traversals.