Changed sphere light sampling to be intentionally slightly incorrect.
The code I was using to calculate the distance of the ray to the sphere's
surface wasn't working correctly.  So for now I'm just stripping it out
and using the distance to the sphere center.  This certainly isn't correct,
but it works fine in most cases, and is better than the previous broken
code.

I've put TODO notes in the code to fix it properly at some point.