Removed unneccesary limit in Tracer code.  Also update TODO.
The limit was a limit placed on the number of potint's that could be
processed at a time.  It was there because there was originally an
additional space cost associated with it.  But since switching to
a partitioning scheme, there is no space cost any longer, and the
limit only serves to prevent amortization of the costs of splitting.