Upon further testing, spheres are better for light clusters than points.
It's important to use the proper solid-angle formula for calculating
light falloff from the sphere, though.  And it's also important to
normalize the brightness of the sphere based on its surface area.  Otherwise
it performs poorly... which makes sense.

Spheres don't perform better on all scenes, but their worst-case performance
is much better, especially with scenes with a large number of un-evenly
distributed lights.