More work on the bare beginnings of a file format.
I've created a simple text-based data tree format, complete with comments.
The parser for this format reads the entire file into memory, and creates
a traversable tree out of it.

Admittedly this is a bit of re-inventing the wheel, as things like
XML, JSON, etc. already exist.  But I wanted something with slightly
different properties.

There is still one major TODO item, which is implementing escape codes
for leaf data and node names.