PsychoBlend now updates viewport after rendering!
It still doesn't update _while_ rendering.  Only after the image is fully
rendered.  But that's still a lot better than before.  This also means that
the images that Blender saves out are actually the rendered images.

Also made it so that in the absence of a defined export path, PsychoBlend
automatically uses temporary files.