Refactored patch/ray intersection code to use a custom stack.
This included creating a type-agnostic Stack class.  The custom stack
is allocated once on the heap for each Tracer, and then re-used by
any intersection routines that need stack-like storage.

This is in preparation for supporting motion blur again, which potentially
needs a large amount of stack space for the iterative intersection
code.