Templatized patch/ray testing, and ripped out DiceableSurface.
- The code for Bilinear and Bicubic patch/ray testing was almost
identical, save for some functions and a data type that had to be
specified for both.  So I've templatized the intersection code, and
just have the Bilinear and Bicubic classes implement the few necessary
functions as static methods.

- It's now very evident that the new iterative approach is faster (and
more accurate!) than the surface dicing approach, so I'm just going for
it and ripping out the DiceableSurface code.  I'm still leaving the
Grid and MicroSurface code in for now, but it's likely to go at some point
in the future as well.