Implemented basic Hero Wavelength spectral sampling.
"Basic" in this case means no spectral MIS, which is half of what
the Hero Wavelength paper is about.  But I don't have any
BSDFs with spectrally dependant directional reflection yet, so
I wouldn't know if the MIS were correct even if I tried to
implement it right now.  So I'm leaving it for the future.

The upshot is that color noise is significantly reduced by this
commit.  At 16 SPP the color noise is now barely noticable
unless you zoom in to pick apart individual pixels.