Working towards multiple importance sampling.
There is now a basic version of MIS more-or-less working, but it
only works with a single light source.  Also, it doesn't appear
to be weighting things _quite_ right even with a single light.
Need to investigate.  But even so, the basic concept seems to be
working, and the renders are close to correct whilst not being
noisey.  So, yay!