Improved SphereLight sampling.
Credit goes to Akalin on ompf2.com for the technique.  See:
http://ompf2.com/viewtopic.php?f=3&t=1914#p4414

Also fixed the bug where tiny sphere lights would cause banding
in the light attenuation.  Basically, I'm making sure that the
important calculations are all being done with doubles.  There
may well be a way to solve this using floats, but for now this
works.