LightTree selection now accounts for the BSDF of the surface.
The basic idea is that each BSDF closure implements its own
method for approximating the contribution of light from a
circular solid angle.  This is then used by the LightTree
to do its importance sampling.

That method is implemented okay for Lambert, but is currently
implemented poorly for GTR.  Both can almost certainly be
improved further.