Stop-gap fix for segfault in subdivision surface rendering.
The issue was that the subdiv bvh was too deep, due to the
pattern of progressively smaller and smaller patches from the
subdiv refinement.  It was getting to over 400 levels deep even
on relatively simple models.  This was causing traversal to
run off the end of the node and ray stacks.

The temporary solution is to slightly tweak how splitting is done
during the BVH build: it takes the mid-point of the bounds of the
patch centroids, rather than the mid-point of the total bounds of
the patches.  However, this doesn't really solve the problem, it
only mitigates it somewhat.  The proper fix, which I'll have to
do in the future, is to implement a more sophisticated splitting
approach.  I'm not quite sure what that will look like.

To help avoid future annoying debugging, the subdiv bvh also
records how deep the tree is, and reports if the tree is too
deep for the traversal code.  This way it will be obvious when
this problem pops up in a scene.