Changed lambert BSDF light tree traversal.
It now uses an accurate analytic calculation of how a sphere light
affects a lambert surface, instead of the weird hack I was using
before.  In some scenes this new approach performs better, but in
others it performs worse.  I've left the hack code in, but
commented, for further investigation at some point.